Added code to encrypt and decrypt using RSA
The javascript part is used to encrypt things and send to a server.

The python code is used to decrypt the data. There is also a small
python (flask) code to receive this data from a local machine.

I also created the public/private keys and they are stored in this same
folder